backend/requirements.txt
pymongo==4.6.3
flask==3.0.2
flask-cors==4.0.1
python-dotenv==1.0.1
pandas==1.3.5
numpy==1.21.4
tqdm==4.62.3